Rename AccountMyPage component to MyPage
컴포넌트 이름을 짧게 변경합니다.
